requirements.txt
opencv-python-headless
numpy
flask
gunicorn
tensorflow-cpu